feat: "side floating navigation" on small screen
We add a "side floating navigation" componenent that appears when the screen gets smaller than a certain size. This component is used to provide clickage round button to navigate to the different sections of the page. We also add a poc that shows how to use Data at Build Time. We only use static data for now (no graphql).