wrap pre-computed SEO component
Attemp to solve an issue in production where the page and metadata are rendred/computed correctly and displayed to the user, but it's only that the console (on browser inspect) shows san error because react re-executes the useStaticQuery hook for fetching site metadata on-client side. The page is correct as mentioned because the useStaticQuery is correclty executed on-server-side too.