feat: make Prof Item contents responsive
Render Professional Item contents as a Fragment of 2 divs, Both divs have CSS property 'display: flex'. Replaces previous 'display: grid' layout. This change makes the Professional Item contents responsive, as the contents will now be displayed in a single column on smaller screens. This is a major change that improves the readability of the Professional Item.