feat: add a macimum width to the side pane
We add CSS properties 'max-width: 400px' and 'width: 100%' to the Div rendered by the VerticalSidePane component. This will ensure that the side pane will not exceed 400px in width, but will take up as much space as possible within that limit. This achieves the design of having the side pane take up smaller width than the main pane when screen is big enough, but also be responsive when viewport shrinks. It practically allows implmenting the ancestor element to either have 'display: flex' or 'display: grid' CSS property.