feat: copy to clipboard on shell command click
Refactored SoftwareReleaseTooltip component to support 2 new props: 'onCopyCommand' a callback with no arguments, and 'copied' a bolean. When 'copied' is true, the tooltip will show a message indicating that the command was copied to the clipboard, which after 2 seconds disappears! The 'onCopyCommand' callback is called when the user clicks on the shell command. So the JSX returned SoftwareReleaseTooltip has been changed a bit to show the "Copied" message. Also, the SoftwareReleaseButtonComponent has been refactored to pass values to the copied and onCopyCommand props, as part of its JSX that instantiates a SoftwareReleaseTooltip component. In this changes there is also a "fix" for SoftwareReleaseButtonComponent, which now does not hide the Release Tooltip when user clicks on either the "Shell Command" or the URL (previously left-clicking anything inside the tooltip would hide it).