feat: run content-aware algo on Resize Item events
Previously, the PortfolioSection Component would add a simple handle to Grid Item onResize Events, that heuristically added 1 or 2 units to the newItem.h based on newItem.w. Now, we bind 3 Refs to the DOM of each Item Content, which allow reporting of offSetWidth, offSetHeight. We then use these values to calculate the new h value for the Item, using an estimation of each Grid Unit as "50px". This is a more accurate way to calculate the h value, and should be more reliable than the previous heuristic, in guaranteeing that each Item Content is maintained as "fully contained" within the Grid Item border.